requirements.txt
praw
python-dotenv
youtube_dl
requests